requirements.txt
pandas==2.3.0
pyspark==4.0.0
python-dotenv==1.1.0
Requests==2.32.4
SQLAlchemy==2.0.41
